rename hints/data to hints/utils
Reset is gonna get pulled into this folder. Data doesn't really need to be its own folder.
